requirements.txt
pandas
matplotlib
scipy
jupyter
scikit-learn
joblib
numpy
hydra-core
dvc
omegaconf
great_expectations
pyyaml
zenml[server]
apache-airflow==2.7.3
feast
pendulum
shyaml
mlflow
giskard==2.14.0
gradio==3.50.2
evidently
Flask
flask-session2
psycopg2-binary
shap
matplotlib
pytest